requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic-settings>=2.0

# LLM & AI
google-genai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Embedding
sentence-transformers>=3.0.0
FlagEmbedding>=1.2.0

# BigQuery
google-cloud-bigquery>=3.25.0
google-cloud-bigquery-storage>=2.25.0

# RAG
docling>=2.0.0
tavily-python>=0.5.0

# Visualization
matplotlib>=3.8.0

# Utilities
structlog>=24.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
locust>=2.30.0
